requirements-dev.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.35.0
anthropic>=0.40.0
notion-client>=2.0.0
playwright>=1.40.0

# PDF processing
PyPDF2>=3.0.0

# Image processing
Pillow>=9.0.0

# Logging and utilities
python-dateutil>=2.8.0
requests>=2.25.0

# Scheduler
apscheduler>=3.10.0
pytz>=2024.1

# Social media integration
tweepy>=4.14.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Cloud storage
boto3>=1.34.0  # For AWS S3

# Optional cloud storage
# cloudinary>=1.40.0  # For Cloudinary

# Chart generation
yfinance>=0.2.28
mplfinance>=0.12.9b7
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0